chore: Move hierarchy builders to correct file
